+++ app/src/plugins/prometheus/PrometheusPage.tsx
@@ -13,6 +13,8 @@ import { useHistory, useLocation } from 'react-router-dom';
 import {
   GetMetricsRequest,
   GetMetricsResponse,
+  MetricLookupRequest,
+  MetricLookupResponse,
   Metrics,
   PrometheusPromiseClient,
   Query,
@@ -80,6 +82,15 @@ const PrometheusPage: React.FunctionComponent<IPluginPageProps> = ({ name, descr
         getMetricsRequest.setResolution(options.resolution);
         getMetricsRequest.setQueriesList(queries);
 
+        const metricLookupRequest = new MetricLookupRequest();
+        metricLookupRequest.setName(name);
+        metricLookupRequest.setMatcher(options.queries[0]);
+        const metricLookupResponse: MetricLookupResponse = await prometheusService.metricLookup(
+          metricLookupRequest,
+          null,
+        );
+        console.log(metricLookupResponse.getNamesList());
+
         const getMetricsResponse: GetMetricsResponse = await prometheusService.getMetrics(getMetricsRequest, null);
         setData({ error: '', isLoading: false, metrics: getMetricsResponse.toObject().metricsList });
       }
@@ -107,6 +118,7 @@ const PrometheusPage: React.FunctionComponent<IPluginPageProps> = ({ name, descr
         </Title>
         <p>{description}</p>
         <PrometheusPageToolbar
+          name={name}
           queries={options.queries}
           resolution={options.resolution}
           timeEnd={options.timeEnd}