+++ app/src/proto/prometheus_pb_service.d.ts
@@ -22,10 +22,20 @@ type PrometheusGetMetrics = {
   readonly responseType: typeof prometheus_pb.GetMetricsResponse;
 };
 
+type PrometheusMetricLookup = {
+  readonly methodName: string;
+  readonly service: typeof Prometheus;
+  readonly requestStream: false;
+  readonly responseStream: false;
+  readonly requestType: typeof prometheus_pb.MetricLookupRequest;
+  readonly responseType: typeof prometheus_pb.MetricLookupResponse;
+};
+
 export class Prometheus {
   static readonly serviceName: string;
   static readonly GetVariables: PrometheusGetVariables;
   static readonly GetMetrics: PrometheusGetMetrics;
+  static readonly MetricLookup: PrometheusMetricLookup;
 }
 
 export type ServiceError = { message: string, code: number; metadata: grpc.Metadata }
@@ -77,6 +87,15 @@ export class PrometheusClient {
   getMetrics(
     requestMessage: prometheus_pb.GetMetricsRequest,
     callback: (error: ServiceError|null, responseMessage: prometheus_pb.GetMetricsResponse|null) => void
+  ): UnaryResponse;
+  metricLookup(
+    requestMessage: prometheus_pb.MetricLookupRequest,
+    metadata: grpc.Metadata,
+    callback: (error: ServiceError|null, responseMessage: prometheus_pb.MetricLookupResponse|null) => void
+  ): UnaryResponse;
+  metricLookup(
+    requestMessage: prometheus_pb.MetricLookupRequest,
+    callback: (error: ServiceError|null, responseMessage: prometheus_pb.MetricLookupResponse|null) => void
   ): UnaryResponse;
 }